requirements.txt
beautifulsoup4==4.14.2
certifi==2025.10.5
charset-normalizer==3.4.4
idna==3.11
requests==2.32.5
soupsieve==2.8
typing_extensions==4.15.0
urllib3==2.5.0
Django~=5.2.7
httpx~=0.28.0
djangorestframework~=3.16.1
pymongo~=4.15.3